Debug character duplicate bug
